AR fixes.
